requirements.txt
streamlit==1.32.0
anthropic==0.25.0
Pillow==10.2.0
pytesseract==0.3.10
pandas==2.2.0
matplotlib==3.8.3
plotly==5.20.0
sqlalchemy==2.0.28
python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.31.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
